Fixed bug: CrlDecimal should produce a decimal, not a double.

+++ Solutions/Menes.JsonSchema.TypeBuilder/Menes.JsonSchema.TypeBuilder/Model/TypeDeclarations.cs
@@ -42,7 +42,7 @@ namespace Menes.JsonSchema.TypeBuilder.Model
         /// <summary>
         /// A clr <see cref="decimal"/> type.
         /// </summary>
-        public static readonly TypeDeclaration ClrDecimalTypeDeclaration = new TypeDeclaration(builtInType: true) { DotnetTypeName = "double" };
+        public static readonly TypeDeclaration ClrDecimalTypeDeclaration = new TypeDeclaration(builtInType: true) { DotnetTypeName = "decimal" };
 
         /// <summary>
         /// A clr <see cref="string"/> type.
